refactor(server): extract workspace provisioning
