+++ server/proliferate/server/cloud/workspaces/provisioning/__init__.py
@@ -1,0 +1,1 @@
+"""Cloud workspace provisioning service package."""
