+++ server/proliferate/server/cloud/workspaces/service.py
@@ -1,97 +1,34 @@
 from __future__ import annotations
 
-from dataclasses import dataclass, replace
 from typing import NoReturn
 from uuid import UUID
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from proliferate.auth.authorization import ActorIdentity, OwnerSelection
-from proliferate.constants.cloud import (
-    SUPPORTED_GIT_PROVIDER,
-    CloudWorkspaceStatus,
-)
-from proliferate.db import session_ops as db_session
-from proliferate.db.models.cloud.workspaces import CloudWorkspace
-from proliferate.db.store import cloud_sandbox_profiles as sandbox_profile_store
-from proliferate.db.store.automation_cloud_workspace_claims import (
-    create_managed_cloud_workspace_for_claimed_run,
-)
-from proliferate.db.store.cloud_agent_auth import store as agent_auth_store
-from proliferate.db.store.cloud_repo_config import (
-    get_cloud_repo_config,
-    get_organization_cloud_repo_config,
-)
 from proliferate.db.store.cloud_workspaces import (
-    CloudRepoLimitExceededError,
-    create_cloud_workspace_for_user,
-    get_existing_managed_cloud_workspace_for_profile,
     list_claimed_organization_workspaces_for_user,
     list_exposed_cloud_workspaces_for_user,
     list_organization_workspaces_for_admin_audit,
     list_unclaimed_organization_workspaces,
-    load_any_cloud_workspace_for_repo,
-    load_existing_cloud_workspace,
-    mark_workspace_error_by_id,
-    save_workspace,
     update_workspace_branch,
     update_workspace_display_name,
 )
 from proliferate.db.store.cloud_workspaces import (
     list_cloud_workspaces as list_cloud_workspaces_store,
 )
-from proliferate.integrations.sandbox import get_configured_sandbox_provider
-from proliferate.server.automations.domain.claim_lifecycle import (
-    CLOUD_WORKSPACE_CREATION_TRANSITION,
-    claim_is_active,
-)
-from proliferate.server.billing.models import SandboxStartAuthorization
-from proliferate.server.billing.service import (
-    authorize_sandbox_start,
-    authorize_sandbox_start_for_billing_subject,
-    get_billing_snapshot_for_subject,
-    repo_limit_for_billing_snapshot,
-)
-from proliferate.server.cloud.agent_auth.domain.status import allowed_agent_kinds
 from proliferate.server.cloud.claims.domain.policy import is_org_admin_role
 from proliferate.server.cloud.errors import CloudApiError
-from proliferate.server.cloud.event_logging import log_cloud_event
-from proliferate.server.cloud.repo_config.service import (
-    bootstrap_repo_config,
-    load_repo_config_value,
-)
-from proliferate.server.cloud.repos.service import (
-    get_linked_github_account,
-)
-from proliferate.server.cloud.repos.service import (
-    get_repo_branches_for_user as get_github_repo_branches,
-)
-from proliferate.server.cloud.runtime.credentials.auth_status import (
-    selected_agent_auth_agent_kinds,
-)
-from proliferate.server.cloud.runtime.scheduler import schedule_workspace_provision
 from proliferate.server.cloud.workspaces.access import (
     cloud_workspace_user_can_interact_with_db,
     cloud_workspace_user_can_read_with_db,
 )
 from proliferate.server.cloud.workspaces.details import (
-    build_workspace_detail as _build_workspace_detail,
-)
-from proliferate.server.cloud.workspaces.details import (
     build_workspace_detail_for_request as _build_workspace_detail_for_request,
-)
-from proliferate.server.cloud.workspaces.details import (
-    cloud_workspace_block_message,
 )
 from proliferate.server.cloud.workspaces.details import (
     workspace_summaries_for_request as _workspace_summaries_for_request,
 )
-from proliferate.server.cloud.workspaces.domain.lifecycle import (
-    decide_workspace_start_after_validation,
-    decide_workspace_status_transition,
-    start_request_should_return_existing,
-)
-from proliferate.server.cloud.workspaces.lifecycle import service as lifecycle_service
 from proliferate.server.cloud.workspaces.models import (
     WorkspaceDetail,
     WorkspaceSummary,
@@ -100,69 +37,12 @@ from proliferate.server.organizations.service import (
     OrganizationServiceError,
     resolve_owner_context,
 )
-from proliferate.utils.time import utcnow
 
 MAX_CLOUD_WORKSPACE_DISPLAY_NAME_CHARS = 160
-CLOUD_HUMAN_ORIGIN_JSON = '{"kind":"human","entrypoint":"cloud"}'
-CLOUD_SYSTEM_ORIGIN_JSON = '{"kind":"system","entrypoint":"cloud"}'
-CREATE_WORKSPACE_ORIGIN_BY_SOURCE = {
-    "desktop": ("manual_desktop", '{"kind":"human","entrypoint":"desktop"}'),
-    "web": ("manual_web", '{"kind":"human","entrypoint":"web"}'),
-    "mobile": ("manual_mobile", '{"kind":"human","entrypoint":"mobile"}'),
-}
-
-
-def _raise_org_cloud_not_ready() -> NoReturn:
-    raise CloudApiError(
-        "org_cloud_not_ready",
-        "Organization cloud workspaces are not available yet.",
-        status_code=409,
-    )
 
 
 def _map_owner_context_error(error: OrganizationServiceError) -> NoReturn:
     raise CloudApiError(error.code, error.message, status_code=error.status_code) from error
-
-
-@dataclass(frozen=True)
-class ResolvedCloudWorkspaceCreate:
-    git_provider: str
-    git_owner: str
-    git_repo_name: str
-    git_branch: str
-    git_base_branch: str
-    display_name: str | None
-    active_sandbox_count: int
-    selected_agent_kinds: tuple[str, ...]
-    cloud_repo_limit: int | None
-
-
-def transition_workspace_status(
-    workspace: CloudWorkspace,
-    target: CloudWorkspaceStatus,
-    *,
-    status_detail: str | None = None,
-) -> None:
-    """Move *workspace* to *target*, enforcing the transition map.
-
-    Raises ``CloudApiError`` when the transition is not allowed.
-    ``status_detail`` overrides the human-readable detail; when *None* the
-    detail is derived from the target status.
-    """
-    decision = decide_workspace_status_transition(
-        workspace.status,
-        target,
-        status_detail=status_detail,
-    )
-    if not decision.allowed:
-        raise CloudApiError(
-            decision.error_code or "invalid_status_transition",
-            decision.error_message or "Invalid workspace status transition.",
-            status_code=decision.status_code or 409,
-        )
-    workspace.status = decision.target_status.value
-    workspace.status_detail = decision.status_detail
-    workspace.updated_at = utcnow()
 
 
 async def list_cloud_workspaces_for_user(
@@ -255,42 +135,6 @@ async def list_cloud_workspaces_for_user(
     return await _workspace_summaries_for_request(db, user_id=user_id, workspaces=workspaces)
 
 
-def _raise_if_cloud_workspace_start_denied(authorization: SandboxStartAuthorization) -> None:
-    if authorization.allowed:
-        return
-    raise CloudApiError(
-        "quota_exceeded",
-        authorization.message or cloud_workspace_block_message(authorization.start_block_reason),
-        status_code=403,
-    )
-
-
-def _raise_repo_limit_exceeded(error: CloudRepoLimitExceededError) -> NoReturn:
-    raise CloudApiError(
-        "repo_limit_exceeded",
-        (
-            "Cloud repo limit reached. Archive an existing cloud repo before adding "
-            f"another one ({error.active_repo_count}/{error.cloud_repo_limit})."
-        ),
-        status_code=403,
-    ) from error
-
-
-async def _load_personal_agent_auth_agent_kinds(user_id: UUID) -> tuple[str, ...]:
-    async with db_session.open_async_session() as db:
-        profile = await agent_auth_store.get_active_personal_sandbox_profile_for_user(
-            db,
-            user_id,
-        )
-        if profile is None:
-            return ()
-        result = await selected_agent_auth_agent_kinds(
-            db,
-            sandbox_profile_id=profile.id,
-        )
-        return result
-
-
 async def get_cloud_workspace_detail(
     db: AsyncSession,
     user_id: UUID,
@@ -298,786 +142,6 @@ async def get_cloud_workspace_detail(
 ) -> WorkspaceDetail:
     workspace = await cloud_workspace_user_can_read_with_db(db, user_id, workspace_id)
     return await _build_workspace_detail_for_request(db, workspace)
-
-
-async def _load_repo_config_value_tx(
-    user_id: UUID, git_owner: str, git_repo_name: str
-) -> object | None:
-    async with db_session.open_async_session() as db:
-        return await load_repo_config_value(
-            db, user_id=user_id, git_owner=git_owner, git_repo_name=git_repo_name
-        )
-
-
-async def _bootstrap_repo_config_tx(user_id: UUID, git_owner: str, git_repo_name: str) -> object:
-    async with db_session.open_async_transaction() as db:
-        return await bootstrap_repo_config(
-            db, user_id=user_id, git_owner=git_owner, git_repo_name=git_repo_name
-        )
-
-
-async def _mark_workspace_error_tx(workspace_id: UUID, message: str, **kwargs: object) -> None:
-    async with db_session.open_async_transaction() as db:
-        await mark_workspace_error_by_id(db, workspace_id, message, **kwargs)
-
-
-async def _resolve_new_cloud_workspace_create(
-    user: ActorIdentity,
-    *,
-    git_provider: str,
-    git_owner: str,
-    git_repo_name: str,
-    base_branch: str | None,
-    branch_name: str,
-    display_name: str | None,
-    required_agent_kind: str | None = None,
-) -> ResolvedCloudWorkspaceCreate:
-    if git_provider != SUPPORTED_GIT_PROVIDER:
-        raise CloudApiError(
-            "unsupported_repo_provider",
-            "Only GitHub repositories are supported for cloud workspaces.",
-            status_code=400,
-        )
-
-    if get_linked_github_account(user) is None:
-        raise CloudApiError(
-            "github_link_required",
-            "Connect a GitHub account before creating a cloud workspace.",
-            status_code=400,
-        )
-
-    cleaned_base_branch = base_branch.strip() if base_branch else ""
-    cleaned_branch_name = branch_name.strip()
-    if not cleaned_branch_name:
-        raise CloudApiError(
-            "invalid_branch_request",
-            "Choose a new cloud branch before creating a cloud workspace.",
-            status_code=400,
-        )
-
-    repo_config = await _load_repo_config_value_tx(user.id, git_owner, git_repo_name)
-    if repo_config is None or not repo_config.configured:
-        async with db_session.open_async_session() as db:
-            existing_repo_workspace = await load_any_cloud_workspace_for_repo(
-                db,
-                user_id=user.id,
-                git_owner=git_owner,
-                git_repo_name=git_repo_name,
-            )
-        if existing_repo_workspace is None:
-            raise CloudApiError(
-                "cloud_repo_not_configured",
-                "Configure cloud settings for this repo before creating a cloud workspace.",
-                status_code=409,
-            )
-        repo_config = await _bootstrap_repo_config_tx(user.id, git_owner, git_repo_name)
-        log_cloud_event(
-            "cloud repo config auto-bootstrapped",
-            user_id=user.id,
-            repo=f"{git_owner}/{git_repo_name}",
-        )
-
-    repo_branches = await get_github_repo_branches(
-        user,
-        git_owner=git_owner,
-        git_repo_name=git_repo_name,
-        missing_access_message="Connect a GitHub account before creating a cloud workspace.",
-        repo_access_required_message=(
-            "Reconnect GitHub and grant repository access before creating a cloud workspace."
-        ),
-    )
-
-    resolved_base_branch = cleaned_base_branch or None
-    saved_default_branch = (repo_config.default_branch or "").strip() or None
-    if resolved_base_branch is None and saved_default_branch:
-        if saved_default_branch in repo_branches.branches:
-            resolved_base_branch = saved_default_branch
-        else:
-            log_cloud_event(
-                "cloud repo default branch missing on github; falling back",
-                user_id=user.id,
-                repo=f"{git_owner}/{git_repo_name}",
-                saved_default_branch=saved_default_branch,
-                github_default_branch=repo_branches.default_branch,
-            )
-    if resolved_base_branch is None:
-        resolved_base_branch = repo_branches.default_branch.strip()
-
-    if resolved_base_branch not in repo_branches.branches:
-        raise CloudApiError(
-            "github_branch_not_found",
-            f"The base branch '{resolved_base_branch}' was not found on GitHub.",
-            status_code=400,
-        )
-    if cleaned_branch_name in repo_branches.branches:
-        raise CloudApiError(
-            "github_branch_already_exists",
-            f"The branch '{cleaned_branch_name}' already exists on GitHub.",
-            status_code=400,
-        )
-
-    async with db_session.open_async_session() as db:
-        existing_cloud_workspace = await load_existing_cloud_workspace(
-            db,
-            user_id=user.id,
-            git_provider=git_provider,
-            git_owner=git_owner,
-            git_repo_name=git_repo_name,
-            git_branch=cleaned_branch_name,
-        )
-    if existing_cloud_workspace is not None:
-        raise CloudApiError(
-            "cloud_branch_already_exists",
-            (
-                f"A cloud workspace already exists for branch '{cleaned_branch_name}'. "
-                "Open the existing cloud workspace or choose a different cloud branch."
-            ),
-            status_code=400,
-        )
-
-    authorization = await authorize_sandbox_start(
-        user_id=user.id,
-        workspace_id=None,
-    )
-    _raise_if_cloud_workspace_start_denied(authorization)
-    billing_snapshot = await get_billing_snapshot_for_subject(authorization.billing_subject_id)
-    cloud_repo_limit = repo_limit_for_billing_snapshot(billing_snapshot)
-
-    if required_agent_kind is not None and required_agent_kind not in allowed_agent_kinds():
-        raise CloudApiError(
-            "unsupported_agent_kind",
-            "The selected agent is not supported for cloud workspaces.",
-            status_code=400,
-        )
-
-    selected_agent_kinds = await _load_personal_agent_auth_agent_kinds(user.id)
-    if required_agent_kind is not None and required_agent_kind not in selected_agent_kinds:
-        raise CloudApiError(
-            "missing_agent_credentials",
-            (
-                f"Select {required_agent_kind} agent authentication before running "
-                "this cloud automation."
-            ),
-            status_code=400,
-        )
-    if not selected_agent_kinds:
-        raise CloudApiError(
-            "missing_supported_credentials",
-            "Select an agent authentication credential before creating a cloud workspace.",
-            status_code=400,
-        )
-    log_cloud_event(
-        "cloud workspace create validated",
-        repo=f"{git_owner}/{git_repo_name}",
-        base_branch=resolved_base_branch,
-        branch_name=cleaned_branch_name,
-        selected_agent_kinds=",".join(selected_agent_kinds),
-        active_sandbox_count=authorization.active_sandbox_count,
-    )
-    return ResolvedCloudWorkspaceCreate(
-        git_provider=git_provider,
-        git_owner=git_owner,
-        git_repo_name=git_repo_name,
-        git_branch=cleaned_branch_name,
-        git_base_branch=resolved_base_branch,
-        display_name=(display_name.strip() if display_name and display_name.strip() else None),
-        active_sandbox_count=authorization.active_sandbox_count,
-        selected_agent_kinds=selected_agent_kinds,
-        cloud_repo_limit=cloud_repo_limit,
-    )
-
-
-async def _resolve_new_managed_cloud_workspace_create(
-    user: ActorIdentity,
-    *,
-    sandbox_profile_id: UUID,
-    target_id: UUID,
-    git_provider: str,
-    git_owner: str,
-    git_repo_name: str,
-    base_branch: str | None,
-    branch_name: str,
-    display_name: str | None,
-    required_agent_kind: str | None = None,
-) -> ResolvedCloudWorkspaceCreate:
-    if git_provider != SUPPORTED_GIT_PROVIDER:
-        raise CloudApiError(
-            "unsupported_repo_provider",
-            "Only GitHub repositories are supported for cloud workspaces.",
-            status_code=400,
-        )
-
-    if get_linked_github_account(user) is None:
-        raise CloudApiError(
-            "github_link_required",
-            "Connect a GitHub account before creating a cloud workspace.",
-            status_code=400,
-        )
-
-    cleaned_base_branch = base_branch.strip() if base_branch else ""
-    cleaned_branch_name = branch_name.strip()
-    if not cleaned_branch_name:
-        raise CloudApiError(
-            "invalid_branch_request",
-            "Choose a new cloud branch before creating a cloud workspace.",
-            status_code=400,
-        )
-
-    async with db_session.open_async_session() as db:
-        profile = await sandbox_profile_store.load_sandbox_profile_by_id(db, sandbox_profile_id)
-        if profile is None or profile.primary_target_id != target_id:
-            raise CloudApiError(
-                "cloud_target_not_found",
-                "Cloud target not found.",
-                status_code=404,
-            )
-        if profile.owner_scope == "organization":
-            if profile.organization_id is None:
-                raise CloudApiError(
-                    "invalid_sandbox_profile",
-                    "Organization sandbox profile is invalid.",
-                    status_code=409,
-                )
-            repo_config = await get_organization_cloud_repo_config(
-                db,
-                organization_id=profile.organization_id,
-                git_owner=git_owner,
-                git_repo_name=git_repo_name,
-            )
-        else:
-            if profile.owner_user_id is None:
-                raise CloudApiError(
-                    "invalid_sandbox_profile",
-                    "Personal sandbox profile is invalid.",
-                    status_code=409,
-                )
-            repo_config = await get_cloud_repo_config(
-                db,
-                user_id=profile.owner_user_id,
-                git_owner=git_owner,
-                git_repo_name=git_repo_name,
-            )
-        if repo_config is None or not repo_config.configured:
-            raise CloudApiError(
-                "cloud_repo_not_configured",
-                "Configure cloud settings for this repo before creating a cloud workspace.",
-                status_code=409,
-            )
-        existing_cloud_workspace = await get_existing_managed_cloud_workspace_for_profile(
-            db,
-            sandbox_profile_id=profile.id,
-            target_id=target_id,
-            git_provider=git_provider,
-            git_owner=git_owner,
-            git_repo_name=git_repo_name,
-            git_branch=cleaned_branch_name,
-        )
-        if existing_cloud_workspace is not None:
-            raise CloudApiError(
-                "cloud_branch_already_exists",
-                (
-                    f"A cloud workspace already exists for branch '{cleaned_branch_name}'. "
-                    "Open the existing cloud workspace or choose a different cloud branch."
-                ),
-                status_code=400,
-            )
-        selected_agent_kinds = await selected_agent_auth_agent_kinds(
-            db,
-            sandbox_profile_id=profile.id,
-        )
-        billing_subject_id = profile.billing_subject_id
-        saved_default_branch = (repo_config.default_branch or "").strip() or None
-        profile_owner_scope = profile.owner_scope
-
-    repo_branches = await get_github_repo_branches(
-        user,
-        git_owner=git_owner,
-        git_repo_name=git_repo_name,
-        missing_access_message="Connect a GitHub account before creating a cloud workspace.",
-        repo_access_required_message=(
-            "Reconnect GitHub and grant repository access before creating a cloud workspace."
-        ),
-    )
-
-    resolved_base_branch = cleaned_base_branch or None
-    if resolved_base_branch is None and saved_default_branch:
-        if saved_default_branch in repo_branches.branches:
-            resolved_base_branch = saved_default_branch
-        else:
-            log_cloud_event(
-                "managed cloud repo default branch missing on github; falling back",
-                user_id=user.id,
-                repo=f"{git_owner}/{git_repo_name}",
-                saved_default_branch=saved_default_branch,
-                github_default_branch=repo_branches.default_branch,
-            )
-    if resolved_base_branch is None:
-        resolved_base_branch = repo_branches.default_branch.strip()
-
-    if resolved_base_branch not in repo_branches.branches:
-        raise CloudApiError(
-            "github_branch_not_found",
-            f"The base branch '{resolved_base_branch}' was not found on GitHub.",
-            status_code=400,
-        )
-    if cleaned_branch_name in repo_branches.branches:
-        raise CloudApiError(
-            "github_branch_already_exists",
-            f"The branch '{cleaned_branch_name}' already exists on GitHub.",
-            status_code=400,
-        )
-
-    authorization = await authorize_sandbox_start_for_billing_subject(
-        actor_user_id=user.id,
-        billing_subject_id=billing_subject_id,
-    )
-    _raise_if_cloud_workspace_start_denied(authorization)
-    billing_snapshot = await get_billing_snapshot_for_subject(authorization.billing_subject_id)
-    cloud_repo_limit = repo_limit_for_billing_snapshot(billing_snapshot)
-
-    if required_agent_kind is not None and required_agent_kind not in allowed_agent_kinds():
-        raise CloudApiError(
-            "unsupported_agent_kind",
-            "The selected agent is not supported for cloud workspaces.",
-            status_code=400,
-        )
-    if profile_owner_scope == "personal":
-        if required_agent_kind is not None and required_agent_kind not in selected_agent_kinds:
-            raise CloudApiError(
-                "missing_agent_credentials",
-                (
-                    f"Select {required_agent_kind} agent authentication before running "
-                    "this cloud automation."
-                ),
-                status_code=400,
-            )
-        if not selected_agent_kinds:
-            raise CloudApiError(
-                "missing_supported_credentials",
-                "Select an agent authentication credential before creating a cloud workspace.",
-                status_code=400,
-            )
-
-    log_cloud_event(
-        "managed cloud workspace create validated",
-        repo=f"{git_owner}/{git_repo_name}",
-        base_branch=resolved_base_branch,
-        branch_name=cleaned_branch_name,
-        selected_agent_kinds=",".join(selected_agent_kinds),
-        active_sandbox_count=authorization.active_sandbox_count,
-        owner_scope=profile_owner_scope,
-        sandbox_profile_id=sandbox_profile_id,
-        target_id=target_id,
-    )
-    return ResolvedCloudWorkspaceCreate(
-        git_provider=git_provider,
-        git_owner=git_owner,
-        git_repo_name=git_repo_name,
-        git_branch=cleaned_branch_name,
-        git_base_branch=resolved_base_branch,
-        display_name=(display_name.strip() if display_name and display_name.strip() else None),
-        active_sandbox_count=authorization.active_sandbox_count,
-        selected_agent_kinds=selected_agent_kinds,
-        cloud_repo_limit=cloud_repo_limit,
-    )
-
-
-async def create_cloud_workspace(
-    user: ActorIdentity,
-    *,
-    db: AsyncSession | None = None,
-    git_provider: str,
-    git_owner: str,
-    git_repo_name: str,
-    base_branch: str | None,
-    branch_name: str,
-    display_name: str | None,
-    required_agent_kind: str | None = None,
-    source: str = "desktop",
-    owner_selection: OwnerSelection | None = None,
-) -> WorkspaceDetail:
-    if owner_selection is not None and owner_selection.owner_scope == "organization":
-        if db is None:
-            raise CloudApiError(
-                "organization_not_found",
-                "Organization not found.",
-                status_code=404,
-            )
-        try:
-            await resolve_owner_context(user, owner_selection, db=db)
-        except OrganizationServiceError as error:
-            _map_owner_context_error(error)
-        _raise_org_cloud_not_ready()
-    resolved = await _resolve_new_cloud_workspace_create(
-        user,
-        git_provider=git_provider,
-        git_owner=git_owner,
-        git_repo_name=git_repo_name,
-        base_branch=base_branch,
-        branch_name=branch_name,
-        display_name=display_name,
-        required_agent_kind=required_agent_kind,
-    )
-
-    try:
-        origin, origin_json = CREATE_WORKSPACE_ORIGIN_BY_SOURCE.get(
-            source,
-            CREATE_WORKSPACE_ORIGIN_BY_SOURCE["desktop"],
-        )
-        async with db_session.open_async_transaction() as create_db:
-            workspace = await create_cloud_workspace_for_user(
-                create_db,
-                user_id=user.id,
-                display_name=resolved.display_name,
-                git_provider=resolved.git_provider,
-                git_owner=resolved.git_owner,
-                git_repo_name=resolved.git_repo_name,
-                git_branch=resolved.git_branch,
-                git_base_branch=resolved.git_base_branch,
-                origin=origin,
-                origin_json=origin_json,
-                template_version=get_configured_sandbox_provider().template_version,
-                cloud_repo_limit=resolved.cloud_repo_limit,
-            )
-    except CloudRepoLimitExceededError as error:
-        _raise_repo_limit_exceeded(error)
-    log_cloud_event(
-        "cloud workspace queued",
-        workspace_id=workspace.id,
-        repo=f"{resolved.git_owner}/{resolved.git_repo_name}",
-        base_branch=resolved.git_base_branch,
-        branch_name=resolved.git_branch,
-    )
-    schedule_workspace_provision(workspace.id)
-    return await _build_workspace_detail(workspace)
-
-
-async def create_cloud_workspace_for_automation_run(
-    user: ActorIdentity,
-    *,
-    run_id: UUID,
-    claim_id: UUID,
-    target_id: UUID | None = None,
-    sandbox_profile_id: UUID | None = None,
-    git_owner: str,
-    git_repo_name: str,
-    branch_name: str,
-    worktree_path: str | None = None,
-    display_name: str | None,
-    required_agent_kind: str,
-) -> CloudWorkspace | None:
-    if target_id is None or sandbox_profile_id is None:
-        raise CloudApiError(
-            "target_required",
-            "Cloud automations require a managed cloud target and sandbox profile.",
-            status_code=409,
-        )
-    resolved = await _resolve_new_managed_cloud_workspace_create(
-        user,
-        sandbox_profile_id=sandbox_profile_id,
-        target_id=target_id,
-        git_provider=SUPPORTED_GIT_PROVIDER,
-        git_owner=git_owner,
-        git_repo_name=git_repo_name,
-        base_branch=None,
-        branch_name=branch_name,
-        display_name=display_name,
-        required_agent_kind=required_agent_kind,
-    )
-    try:
-        async with db_session.open_async_transaction() as db:
-            workspace = await create_managed_cloud_workspace_for_claimed_run(
-                db,
-                run_id=run_id,
-                claim_id=claim_id,
-                sandbox_profile_id=sandbox_profile_id,
-                target_id=target_id,
-                user_id=user.id,
-                display_name=resolved.display_name,
-                git_provider=resolved.git_provider,
-                git_owner=resolved.git_owner,
-                git_repo_name=resolved.git_repo_name,
-                git_branch=resolved.git_branch,
-                git_base_branch=resolved.git_base_branch,
-                worktree_path=worktree_path,
-                origin_json=CLOUD_SYSTEM_ORIGIN_JSON,
-                template_version=get_configured_sandbox_provider().template_version,
-                now=utcnow(),
-                transition=CLOUD_WORKSPACE_CREATION_TRANSITION,
-                claim_is_active=claim_is_active,
-            )
-    except CloudRepoLimitExceededError as error:
-        _raise_repo_limit_exceeded(error)
-    if workspace is not None:
-        log_cloud_event(
-            "automation cloud workspace created",
-            workspace_id=workspace.id,
-            automation_run_id=run_id,
-            repo=f"{resolved.git_owner}/{resolved.git_repo_name}",
-            branch_name=resolved.git_branch,
-        )
-    return workspace
-
-
-async def ensure_cloud_workspace_for_existing_branch(
-    user: ActorIdentity,
-    *,
-    git_provider: str,
-    git_owner: str,
-    git_repo_name: str,
-    branch_name: str,
-    display_name: str | None,
-) -> CloudWorkspace:
-    if git_provider != SUPPORTED_GIT_PROVIDER:
-        raise CloudApiError(
-            "unsupported_repo_provider",
-            "Only GitHub repositories are supported for cloud workspaces.",
-            status_code=400,
-        )
-    if get_linked_github_account(user) is None:
-        raise CloudApiError(
-            "github_link_required",
-            "Connect a GitHub account before provisioning a cloud workspace.",
-            status_code=400,
-        )
-    cleaned_branch_name = branch_name.strip()
-    if not cleaned_branch_name:
-        raise CloudApiError(
-            "invalid_branch_request",
-            "Choose a branch before provisioning a cloud workspace.",
-            status_code=400,
-        )
-    async with db_session.open_async_session() as db:
-        existing_cloud_workspace = await load_existing_cloud_workspace(
-            db,
-            user_id=user.id,
-            git_provider=git_provider,
-            git_owner=git_owner,
-            git_repo_name=git_repo_name,
-            git_branch=cleaned_branch_name,
-        )
-    if existing_cloud_workspace is not None:
-        if _cloud_workspace_should_be_replaced_for_mobility_retry(existing_cloud_workspace):
-            await lifecycle_service.archive_failed_cloud_workspace_for_mobility_retry(
-                existing_cloud_workspace.id
-            )
-            log_cloud_event(
-                "failed cloud workspace archived before mobility retry",
-                workspace_id=existing_cloud_workspace.id,
-                repo=f"{git_owner}/{git_repo_name}",
-                branch_name=cleaned_branch_name,
-                last_error=existing_cloud_workspace.last_error,
-            )
-        else:
-            return existing_cloud_workspace
-    repo_config = await _load_repo_config_value_tx(user.id, git_owner, git_repo_name)
-    if repo_config is None or not repo_config.configured:
-        repo_config = await _bootstrap_repo_config_tx(user.id, git_owner, git_repo_name)
-        log_cloud_event(
-            "cloud repo config auto-bootstrapped",
-            user_id=user.id,
-            repo=f"{git_owner}/{git_repo_name}",
-        )
-    authorization = await authorize_sandbox_start(
-        user_id=user.id,
-        workspace_id=None,
-    )
-    _raise_if_cloud_workspace_start_denied(authorization)
-    billing_snapshot = await get_billing_snapshot_for_subject(authorization.billing_subject_id)
-    cloud_repo_limit = repo_limit_for_billing_snapshot(billing_snapshot)
-    try:
-        async with db_session.open_async_transaction() as db:
-            workspace = await create_cloud_workspace_for_user(
-                db,
-                user_id=user.id,
-                display_name=display_name.strip()
-                if display_name and display_name.strip()
-                else None,
-                git_provider=git_provider,
-                git_owner=git_owner,
-                git_repo_name=git_repo_name,
-                git_branch=cleaned_branch_name,
-                git_base_branch=cleaned_branch_name,
-                origin_json=CLOUD_HUMAN_ORIGIN_JSON,
-                template_version=get_configured_sandbox_provider().template_version,
-                cloud_repo_limit=cloud_repo_limit,
-            )
-    except CloudRepoLimitExceededError as error:
-        _raise_repo_limit_exceeded(error)
-    log_cloud_event(
-        "cloud workspace ensured for mobility",
-        workspace_id=workspace.id,
-        repo=f"{git_owner}/{git_repo_name}",
-        branch_name=cleaned_branch_name,
-    )
-    return workspace
-
-
-def _cloud_workspace_should_be_replaced_for_mobility_retry(
-    workspace: CloudWorkspace,
-) -> bool:
-    return (
-        workspace.status == CloudWorkspaceStatus.error.value
-        and workspace.anyharness_workspace_id is None
-    )
-
-
-async def _refresh_repo_env_snapshot_for_workspace(workspace: CloudWorkspace) -> CloudWorkspace:
-    repo_config = await _load_repo_config_value_tx(
-        workspace.user_id, workspace.git_owner, workspace.git_repo_name
-    )
-    if repo_config is None or not repo_config.configured:
-        repo_config = await _bootstrap_repo_config_tx(
-            workspace.user_id, workspace.git_owner, workspace.git_repo_name
-        )
-        log_cloud_event(
-            "cloud repo config auto-bootstrapped",
-            user_id=workspace.user_id,
-            repo=f"{workspace.git_owner}/{workspace.git_repo_name}",
-        )
-    async with db_session.open_async_transaction() as db:
-        return await save_workspace(db, workspace)
-
-
-async def start_cloud_workspace(
-    db: AsyncSession,
-    user: ActorIdentity,
-    workspace_id: UUID,
-    *,
-    requested_base_sha: str | None = None,
-) -> WorkspaceDetail:
-    workspace = await cloud_workspace_user_can_interact_with_db(db, user.id, workspace_id)
-    has_requested_revision = bool((requested_base_sha or "").strip())
-    if start_request_should_return_existing(workspace.status) and not has_requested_revision:
-        return await _build_workspace_detail(workspace)
-    if has_requested_revision and workspace.status == CloudWorkspaceStatus.materializing.value:
-        raise CloudApiError(
-            "cloud_workspace_already_materializing",
-            "Cloud workspace is already preparing. Try the move again once it is ready.",
-            status_code=409,
-        )
-
-    repo_branches = await get_github_repo_branches(
-        user,
-        git_owner=workspace.git_owner,
-        git_repo_name=workspace.git_repo_name,
-        missing_access_message="Connect a GitHub account before starting a cloud workspace.",
-        repo_access_required_message=(
-            "Reconnect GitHub and grant repository access before starting a cloud workspace."
-        ),
-    )
-    base_branch = workspace.git_base_branch or workspace.git_branch
-    if base_branch not in repo_branches.branches:
-        raise CloudApiError(
-            "github_branch_not_found",
-            f"The base branch '{base_branch}' was not found on GitHub.",
-            status_code=400,
-        )
-
-    authorization = await authorize_sandbox_start(
-        user_id=user.id,
-        workspace_id=workspace.id,
-    )
-    _raise_if_cloud_workspace_start_denied(authorization)
-
-    selected_agent_kinds = await _load_personal_agent_auth_agent_kinds(user.id)
-    if not selected_agent_kinds:
-        raise CloudApiError(
-            "missing_supported_credentials",
-            "Select an agent authentication credential before starting a cloud workspace.",
-            status_code=400,
-        )
-    log_cloud_event(
-        "cloud workspace start validated",
-        workspace_id=workspace.id,
-        repo=f"{workspace.git_owner}/{workspace.git_repo_name}",
-        base_branch=base_branch,
-        branch_name=workspace.git_branch,
-        selected_agent_kinds=",".join(selected_agent_kinds),
-        active_sandbox_count=authorization.active_sandbox_count,
-    )
-
-    start_decision = decide_workspace_start_after_validation(
-        workspace.status,
-        ready_at_exists=workspace.ready_at is not None,
-    )
-    if start_decision.refresh_repo_env_snapshot:
-        workspace = await _refresh_repo_env_snapshot_for_workspace(workspace)
-        start_decision = decide_workspace_start_after_validation(
-            workspace.status,
-            ready_at_exists=workspace.ready_at is not None,
-        )
-
-    if start_decision.action == "queue_pending":
-        if start_decision.clear_last_error:
-            workspace.last_error = None
-        if start_decision.persist_before_schedule:
-            async with db_session.open_async_transaction() as persist_db:
-                workspace = await save_workspace(persist_db, workspace)
-        log_cloud_event(
-            "cloud workspace queued",
-            workspace_id=workspace.id,
-            repo=f"{workspace.git_owner}/{workspace.git_repo_name}",
-            base_branch=base_branch,
-            branch_name=workspace.git_branch,
-            requested_base_sha=requested_base_sha,
-        )
-        if start_decision.schedule_provision:
-            schedule_workspace_provision(
-                workspace.id,
-                requested_base_sha=requested_base_sha,
-            )
-        return await _build_workspace_detail(workspace)
-
-    if start_decision.action in {"return_ready", "return_current"}:
-        if not has_requested_revision:
-            return await _build_workspace_detail(workspace)
-        if start_decision.action == "return_current":
-            raise CloudApiError(
-                "cloud_workspace_already_materializing",
-                "Cloud workspace is already preparing. Try the move again once it is ready.",
-                status_code=409,
-            )
-        start_decision = replace(
-            start_decision,
-            action="restart_materializing",
-            clear_last_error=True,
-            persist_before_schedule=True,
-            schedule_provision=True,
-            target_status=CloudWorkspaceStatus.materializing,
-            status_detail="Preparing requested revision",
-        )
-
-    if start_decision.target_status is None:
-        raise CloudApiError(
-            "invalid_status_transition",
-            "Invalid workspace status transition.",
-            status_code=409,
-        )
-    transition_workspace_status(
-        workspace,
-        start_decision.target_status,
-        status_detail=start_decision.status_detail,
-    )
-    if start_decision.clear_last_error:
-        workspace.last_error = None
-    if start_decision.persist_before_schedule:
-        async with db_session.open_async_transaction() as persist_db:
-            workspace = await save_workspace(persist_db, workspace)
-    log_cloud_event(
-        "cloud workspace restart queued",
-        workspace_id=workspace.id,
-        repo=f"{workspace.git_owner}/{workspace.git_repo_name}",
-        base_branch=base_branch,
-        branch_name=workspace.git_branch,
-        requested_base_sha=requested_base_sha,
-    )
-    if start_decision.schedule_provision:
-        schedule_workspace_provision(
-            workspace.id,
-            requested_base_sha=requested_base_sha,
-        )
-    return await _build_workspace_detail(workspace)
 
 
 async def sync_cloud_workspace_branch(
